Updated test_utils to include a new test for contrast enhancer

+++ tests/test_utils.py
@@ -916,8 +916,8 @@ def test_contrast_enhancer() -> None:
     # The out_put array should be equal to expected result_array
     assert np.all(result_array == output_array)
     # Calculating the contrast enhanced version of input_array
-    input_array = np.array([0, 255])
-    result_array = np.array([0, 255])
+    input_array = np.array([0, 255], dtype=np.uint8)
+    result_array = np.array([0, 255], dtype=np.uint8)
     output_array = utils.misc.contrast_enhancer(input_array, low_p=98, high_p=2)
     # The out_put array should be equal to expected result_array
     assert np.all(result_array == output_array)
